+++ p2panda-rs/src/storage_provider/traits/mod.rs
@@ -15,7 +15,7 @@ mod test_utils;
 pub use document_store::DocumentStore;
 pub use entry_store::EntryStore;
 pub use log_store::LogStore;
-pub use models::{AsStorageEntry, AsStorageLog, AsStorageOperation};
+pub use models::{AsStorageEntry, AsStorageLog};
 pub use operation_store::OperationStore;
 pub use requests::{AsEntryArgsRequest, AsPublishEntryRequest};
 pub use responses::{AsEntryArgsResponse, AsPublishEntryResponse};